Create TelosOperatingAgreement
The Telos Network Operating Agreement is the Telos version of the EOS constitution which describes the operating rules for the network. Review, discussion, and changes are expected in this document. The github version of this document will always be the most current until it is accepted as a final draft and recorded on the Telos blockchain.